+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [4.3.1]
+### Added
+- Flag generateEmbedding to mark profiles that should generate embeddings
+### Changed in DB or environment
+- Automated endorser partner SQL script 6
+
 ## [4.3.0]
 ### Added
 - `/api/partner/userProfileNearestNeighbors/:rowId` endpoint to find nearest connection in registration tree to a user profile